ファイル名の変更とMTGの結果やらないことにした変更点の切り戻し
