+++ src/main/java/highandlow_answer/HighAndLow1ChangeDefault.java
@@ -9,7 +9,7 @@ import java.util.regex.Pattern;
 /**
  * High and Low ゲームクラス
  */
-class HighAndLow_2_tryWithResources {
+class HighAndLow1ChangeDefault {
     public static void main(String[] args) {
         // スタート文言の表示
         System.out.println("High and Low ゲームをはじめます。");
@@ -49,8 +49,11 @@ class HighAndLow_2_tryWithResources {
 
         String userAnswer = "";
 
-        try(InputStreamReader is = new InputStreamReader(System.in);
-            BufferedReader br = new BufferedReader(is)) {
+        InputStreamReader is = null;
+        BufferedReader br = null;
+        try {
+            is = new InputStreamReader(System.in);
+            br = new BufferedReader(is);
 
             // ユーザの入力を待ちます
             userAnswer = br.readLine();
@@ -64,6 +67,14 @@ class HighAndLow_2_tryWithResources {
 
         } catch (IOException ioe) {
             System.out.println("入力値に誤りがあります。ゲームを始めからやりなおしてください。 入力値：" + userAnswer);
+
+        } finally {
+            try {
+                is.close();
+                br.close();
+            } catch (IOException ioe) {
+                System.out.println("クローズ処理に失敗しました。");
+            }
         }
 
         return Integer.parseInt(userAnswer);
